CLIPMulti_reduced_project/CLIPMulti_reduced/requirements.txt

torch
torchvision
transformers
sentencepiece
Pillow
tqdm
scikit-learn
pandas
